replace short with int

+++ 004/sol.cpp
@@ -13,12 +13,12 @@ bool is_palindrome(int number) {
 }
 
 int main() {
-    const short n = 999;
+    const int n = 999;
     int mul = 0;
     int res = 0;
     
-    for (short i = 1; i <= n; i++) {
-        for (short j = 1; j <= n; j++) {
+    for (int i = 1; i <= n; i++) {
+        for (int j = 1; j <= n; j++) {
             mul = i * j;
             if (is_palindrome(mul) && mul > res) {
                 res = mul;
